+++ packages/database/src/database/models/__init__.py
@@ -1,5 +1,6 @@
 from database.models.agent_run import AgentRun, ModelCall
 from database.models.document import Document, DocumentChunk
+from database.models.graph import DBKnowledgeEntity, DBKnowledgeRelation
 from database.models.memory import DBResearchMemory
 from database.models.report import Report
 from database.models.research_job import ResearchJob, ResearchTask
@@ -8,8 +9,10 @@ from database.models.usage_record import UsageRecord
 from database.models.user import User
 from database.models.user_quota import UserQuota
 
-# Alias for standard naming
+# Aliases for standard naming
 ResearchMemory = DBResearchMemory
+KnowledgeEntity = DBKnowledgeEntity
+KnowledgeRelation = DBKnowledgeRelation
 
 __all__ = [
     "ResearchJob",
@@ -26,4 +29,8 @@ __all__ = [
     "UserQuota",
     "DBResearchMemory",
     "ResearchMemory",
+    "DBKnowledgeEntity",
+    "DBKnowledgeRelation",
+    "KnowledgeEntity",
+    "KnowledgeRelation",
 ]